+++ indispensable-asn1/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1String.kt
@@ -1,86 +1,349 @@
 package at.asitplus.signum.indispensable.asn1
 
-import at.asitplus.signum.indispensable.asn1.encoding.asAsn1String
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
+import at.asitplus.signum.indispensable.asn1.BERTags.BMP_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.GENERAL_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.GRAPHIC_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.IA5_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.NUMERIC_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.PRINTABLE_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.T61_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.UNIVERSAL_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.UNRESTRICTED_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.UTF8_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.VIDEOTEX_STRING
+import at.asitplus.signum.indispensable.asn1.BERTags.VISIBLE_STRING
+import at.asitplus.signum.indispensable.asn1.encoding.decodeFromAsn1ContentBytes
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToBmpString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToGeneralString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToGraphicString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToIa5String
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToNumericString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToPrintableString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToTeletextString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToUniversalString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToUnrestrictedString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToUtf8String
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToVideotexString
+import at.asitplus.signum.indispensable.asn1.encoding.decodeToVisibleString
 
-//TODO auto-sanitize and/or reduce
+
 /**
- * ASN.! String class used as wrapper do discriminate between different ASN.1 string types
+ * ASN.1 String class used as wrapper do discriminate between different ASN.1 string types
+ * By default, the string value is decoded using UTF-8. If a different charset or custom decoding
+ * is needed, the [rawValue] property can be used directly.
+ *
+ * To enable parsing of non-compliant strings without exploding, every String is internally represented
+ * as raw [ByteArray] and not validated during decoding from ASN.1.
+ * The [isValid] property indicates whether the bytes contained in an ASN.1 String object type are valid
+ * according to the validation rules of that type.
  */
-sealed class Asn1String : Asn1Encodable<Asn1Primitive> {
+sealed class Asn1String(
+    val rawValue: ByteArray,
+    val performValidation: Boolean
+) : Asn1Encodable<Asn1Primitive> {
     abstract val tag: ULong
-    abstract val value: String
+
+    /**
+     * Always the UTF-8 interpretation of [rawValue].
+     * The decoding is performed via [String.decodeFromAsn1ContentBytes], which internally uses
+     * the standard library's [ByteArray.decodeToString].
+     */
+    val value: String by lazy { String.decodeFromAsn1ContentBytes(rawValue) }
+
+    /**
+     * Returns whether this string is valid:
+     * - `true`: validation succeeded
+     * - `false`: validation failed
+     * - `null`: no validation implemented
+     */
+    abstract val isValid: Boolean?
+
 
     /**
      * UTF8 STRING (verbatim String)
      */
-    class UTF8(override val value: String) : Asn1String() {
+    class UTF8 private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.UTF8_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            !value.contains('\uFFFD')
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
-     * UNIVERSAL STRING (unchecked)
+     * UNIVERSAL STRING (no checks)
+     * Validation is not implemented. This string format is deprecated for HTTPS certificates and its use in generally discouraged in favor of UTF-8 strings (see [Asn1String.UTF8]).
      */
-    class Universal(override val value: String) : Asn1String() {
+    class Universal private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.UNIVERSAL_STRING.toULong()
+
+        /**
+         * Always `null`, since no validation logic is implemented
+         */
+        override val isValid: Boolean? = null
+
+        constructor(value: String) : this(value.encodeToByteArray(), false)
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
-     * VISIBLE STRING (no checks)
+     * VISIBLE STRING (checked)
      */
-    class Visible(override val value: String) : Asn1String() {
+    class Visible private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.VISIBLE_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[\\x20-\\x7E]*").matches(value)
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
-     * IA5 STRING (no checks)
+     * IA5 STRING (checked)
      */
-    class IA5(override val value: String) : Asn1String() {
+    class IA5 private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.IA5_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[\\x00-\\x7E]*").matches(value)
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
-     * TELETEX STRING (no checks)
+     * TELETEX STRING (checked).
+     *  This string format is deprecated for HTTPS certificates and its use in generally discouraged in favor of UTF-8 strings (see [Asn1String.UTF8]).
      */
-    class Teletex(override val value: String) : Asn1String() {
+    class Teletex private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.T61_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[\\u0000-\\u00FF]*").matches(value)
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
-     * BMP STRING (no checks)
+     * BMP STRING (unchecked).
+     * Validation is not implemented. This string format is deprecated for HTTPS certificates and its use in generally discouraged in favor of UTF-8 strings (see [Asn1String.UTF8]).
      */
-    class BMP(override val value: String) : Asn1String() {
+    class BMP private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
         override val tag = BERTags.BMP_STRING.toULong()
+
+        /**
+         * Always `null`, since no validation logic is implemented
+         */
+        override val isValid: Boolean? = null
+
+        constructor(value: String) : this(value.encodeToByteArray(), false)
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
+    }
+
+    /**
+     * GENERAL STRING (checked)
+     */
+    class General private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.GENERAL_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[\\x00-\\x7E]*").matches(value)
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
+    }
+
+    /**
+     * GRAPHIC STRING (checked)
+     */
+    class Graphic private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.GRAPHIC_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[\\x20-\\x7E]*").matches(value)
+        }
+
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
+    }
+
+    /**
+     * CHARACTER/UNRESTRICTED STRING (no checks)
+     */
+    class Unrestricted private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.UNRESTRICTED_STRING.toULong()
+
+        /**
+         * Always `null`, since no validation logic is implemented
+         */
+        override val isValid: Boolean? = null
+
+        constructor(value: String) : this(value.encodeToByteArray(), false)
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
+    }
+
+    /**
+     * VIDEOTEX STRING (no checks)
+     * Validation is not implemented. This type is no longer used in practice.
+     */
+    class Videotex private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.VIDEOTEX_STRING.toULong()
+
+        /**
+         * Always `null`, since no validation logic is implemented
+         */
+        override val isValid: Boolean? = null
+
+        constructor(value: String) : this(value.encodeToByteArray(), false)
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
      * PRINTABLE STRING (checked)
-     * @throws Asn1Exception if illegal characters are provided
      */
-    class Printable @Throws(Asn1Exception::class) constructor(override val value: String) : Asn1String() {
-        init {
-            Regex("[a-zA-Z0-9 '()+,-./:=?]*").matchEntire(value)
-                ?: throw Asn1Exception("Input contains invalid chars: '$value'")
+    class Printable private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.PRINTABLE_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[a-zA-Z0-9 '()+,-./:=?]*").matches(value)
         }
 
-        override val tag = BERTags.PRINTABLE_STRING.toULong()
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
     /**
      * NUMERIC STRING (checked)
-     * @throws Asn1Exception if illegal characters are provided
      */
-    class Numeric @Throws(Asn1Exception::class) constructor(override val value: String) : Asn1String() {
-        init {
-            Regex("[0-9 ]*").matchEntire(value)
-                ?: throw Asn1Exception("Input contains invalid chars: '$value'")
+    class Numeric private constructor(
+        rawValue: ByteArray,
+        performValidation: Boolean
+    ) : Asn1String(rawValue, performValidation) {
+        override val tag = BERTags.NUMERIC_STRING.toULong()
+
+        override val isValid: Boolean by lazy {
+            Regex("[0-9 ]*").matches(value)
         }
 
-        override val tag = BERTags.NUMERIC_STRING.toULong()
+        /**
+         * @throws Asn1Exception if illegal characters are provided
+         */
+        @Throws(Asn1Exception::class)
+        constructor(value: String) : this(value.encodeToByteArray(), true) {
+            if (!isValid) throw Asn1Exception("Input contains invalid chars: '$value'")
+        }
+
+        @PublishedApi
+        internal constructor(rawValue: ByteArray) : this(rawValue, false)
     }
 
-    override fun encodeToTlv() = Asn1Primitive(tag, value.encodeToByteArray())
+    override fun encodeToTlv() = Asn1Primitive(tag, rawValue)
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other == null || this::class != other::class) return false
@@ -101,7 +364,33 @@ sealed class Asn1String : Asn1Encodable<Asn1Primitive> {
 
     companion object : Asn1Decodable<Asn1Primitive, Asn1String> {
 
+        /**
+         * Decodes an [Asn1Primitive] into a specific [Asn1String] subtype based on its tag.
+         *
+         * For cases where an implicit tag is required, see the helper extension methods
+         * like [decodeToUtf8String], [decodeToPrintableString], etc., which allow specifying an optional tag override.
+         *
+         * @param src the ASN.1 primitive to decode
+         * @return the corresponding [Asn1String] subtype
+         * @throws Asn1Exception if decoding fails or the tag is unsupported
+         */
         @Throws(Asn1Exception::class)
-        override fun doDecode(src: Asn1Primitive): Asn1String = src.asAsn1String()
+        override fun doDecode(src: Asn1Primitive): Asn1String = runRethrowing {
+            when (src.tag.tagValue) {
+                UTF8_STRING.toULong() -> src.decodeToUtf8String()
+                UNIVERSAL_STRING.toULong() -> src.decodeToUniversalString()
+                IA5_STRING.toULong() -> src.decodeToIa5String()
+                BMP_STRING.toULong() -> src.decodeToBmpString()
+                T61_STRING.toULong() -> src.decodeToTeletextString()
+                PRINTABLE_STRING.toULong() -> src.decodeToPrintableString()
+                NUMERIC_STRING.toULong() -> src.decodeToNumericString()
+                VISIBLE_STRING.toULong() -> src.decodeToVisibleString()
+                GENERAL_STRING.toULong() -> src.decodeToGeneralString()
+                GRAPHIC_STRING.toULong() -> src.decodeToGraphicString()
+                UNRESTRICTED_STRING.toULong() -> src.decodeToUnrestrictedString()
+                VIDEOTEX_STRING.toULong() -> src.decodeToVideotexString()
+                else -> throw Asn1Exception("Not an Asn1String!")
+            }
+        }
     }
 }